Moved codegen into its own file. Added unsafe calls to llvm bindings.
