+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod syntax;
+pub mod llvm;
